middle point of transition from :master API to :unstable API

+++ src/internal.c
@@ -69,7 +69,7 @@ int send_cmd(freenect_device *dev, uint16_t cmd, void *cmd_buf, unsigned int cmd
 		actual_len = fnusb_control(&dev->usb_cam, 0xc0, 0, 0, 0, ibuf, 0x200);
 	} while (actual_len == 0);
 	FN_SPEW("Control reply: %d\n", res);
-	if (actual_len < sizeof(*rhdr)) {
+	if (actual_len < (int)sizeof(*rhdr)) {
 		FN_ERROR("send_cmd: Input control transfer failed (%d)\n", res);
 		return res;
 	}
